Fix the order of columns in column export.
The order in assays in column exports (export of
assay data, as in the SAMPLE and TFAM formats),
should correspond to the assay order as used in the
high dimensional data exports.
A new function has been added to high dimensional
data types that retrieves assay columns in the correct
order.